+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Support direct-access mode on AQT devices (#164)
 * Support up to 2000 shots per circuit on cloud resources (#165)
 * Fix result timeout propagation in direct-access mode (#171)
+* Don't propagate empty access tokens in direct-access mode (#172)
 
 ## qiskit-aqt-provider v1.5.0