+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## qiskit-aqt-provider v1.6.0
 
 * Support direct-access mode on AQT devices (#164)
 * Support up to 2000 shots per circuit on cloud resources (#165)